controllers changes

+++ app/controllers/comments_controller.rb
@@ -1,5 +1,5 @@
 class CommentsController < ApplicationController
-	before_action :load_commentable , only: [:index, :show]
+	before_action :load_commentable
 	before_action :authenticate_user!, only: [:create, :update, :destroy]
 
 	def index
@@ -13,8 +13,9 @@ class CommentsController < ApplicationController
 	end
 
 	def create
-		user = current_user
-		comment = current_user.comments.create(comment_params)
+		pa = comment_params
+		pa[:comment].merge!(:user_id => current_user.id)
+		comment = @commentable.comments.create(pa)
 		if comment.valid?
 			render json: comment, status: :created
 		else
@@ -24,43 +25,43 @@ class CommentsController < ApplicationController
 
 	
 	def update
-		if @commentable
-			comment = @commentable.comments.find(params[:id])
-		else 
-			comment = Comment.find(params[:id])
+		pa = comment_params
+		pa[:comment].merge!(:user_id => current_user.id)
+		comment = @commentable.comments.find(params[:id])
+		if comment.user_id == current_user.id
+			if comment.update(pa)
+				render json: comment, status: :ok
+			else
+				render json: comment.errors, status: :unprocessable_entity
+			end
+		else
+			render json: { error: "can't find commentable for the given user"}, status: :not_found
 		end
 		
-		if comment.update(comment_params)
-			render json: comment, status: :ok
-		else
-			render json: comment.errors, status: :unprocessable_entity
-		end
 	end
 	
 	def destroy
-		if @commentable
-			comment = @commentable.coments.find(params[:id])
+		comment = @commentable.comments.find(params[:id])
+		if comment.user_id == current_user.id
+			comment.destroy
+			if comment.destroyed?
+				render json: comment, status: :ok
+			else
+				render json: comment.errors, status: :unprocessable_entity
+			end
 		else
-			comment = Comment.find(params[:id])
-		end
-		
-		comment.destroy
-		if comment.destroyed?
-			render json: comment, status: :ok
-		else
-			render json: comment.errors, status: :unprocessable_entity
+			render json: { error: "can't find commentable for the given user"}, status: :not_found
 		end
 	end
 
 	private 
 	def load_commentable
-		if(user_signed_in?)
 		resource, id = request.path.split('/')[1,2]
 		@commentable = resource.singularize.classify.constantize.find(id)
 	end
 
 	def comment_params
-		params.require(:comment).permit(:user_id, :body)
+		params.require(:comment).permit(:body)
 	end
 
 end
